[FEAT] Add save maze action
